+++ compose/src/commonMain/kotlin/com/blockstream/compose/navigation/Router.kt
@@ -27,6 +27,8 @@ import com.blockstream.common.models.add.XpubViewModel
 import com.blockstream.common.models.addresses.AddressesViewModel
 import com.blockstream.common.models.addresses.SignMessageViewModel
 import com.blockstream.common.models.archived.ArchivedAccountsViewModel
+import com.blockstream.common.models.assetaccounts.AssetAccountDetailsViewModel
+import com.blockstream.common.models.assetaccounts.AssetAccountListViewModel
 import com.blockstream.common.models.camera.CameraViewModel
 import com.blockstream.common.models.devices.DeviceInfoViewModel
 import com.blockstream.common.models.devices.DeviceListViewModel
@@ -104,6 +106,8 @@ import com.blockstream.compose.screens.add.ReviewAddAccountScreen
 import com.blockstream.compose.screens.add.XpubScreen
 import com.blockstream.compose.screens.addresses.AddressesScreen
 import com.blockstream.compose.screens.archived.ArchivedAccountsScreen
+import com.blockstream.compose.screens.assetaccounts.AssetAccountDetailsScreen
+import com.blockstream.compose.screens.assetaccounts.AssetAccountListScreen
 import com.blockstream.compose.screens.devices.DeviceInfoScreen
 import com.blockstream.compose.screens.devices.DeviceListScreen
 import com.blockstream.compose.screens.devices.DeviceScanScreen
@@ -450,7 +454,8 @@ fun Router(
                     SendViewModel(
                         greenWallet = args.greenWallet,
                         initAddress = args.address,
-                        addressType = args.addressType
+                        addressType = args.addressType,
+                        initialAccountAsset = args.accountAsset
                     )
                 })
             }
@@ -872,6 +877,28 @@ fun Router(
                         )
                     },
                     onDismissRequest = navController.onDismissRequest()
+                )
+            }
+            appComposable<NavigateDestinations.AssetAccountList> {
+                val args = it.toRoute<NavigateDestinations.AssetAccountList>()
+                AssetAccountListScreen(
+                    viewModel = viewModel {
+                        AssetAccountListViewModel(
+                            greenWallet = args.greenWallet,
+                            assetId = args.assetId
+                        )
+                    }
+                )
+            }
+            appComposable<NavigateDestinations.AssetAccountDetails> {
+                val args = it.toRoute<NavigateDestinations.AssetAccountDetails>()
+                AssetAccountDetailsScreen(
+                    viewModel = viewModel {
+                        AssetAccountDetailsViewModel(
+                            greenWallet = args.greenWallet,
+                            accountAssetOrNull = args.accountAsset
+                        )
+                    }
                 )
             }
             appBottomSheet<NavigateDestinations.JadeFirmwareUpdate> {